Fixed camera yaw bug.

+++ platform_tilt_brute_forcer/platform_tilt_brute_forcer/main_hau.cpp
@@ -1603,7 +1603,7 @@ __global__ void check_speed_angle() {
 
                         for (int k = 1; k < 4; k++) {
                             int cameraYaw = calculate_camera_yaw(oneUpPlatformPosition, cameraPositions[k]);
-                            cameraYaw = cameraYaw - refCameraYaw;
+                            cameraYaw = (short)(cameraYaw - refCameraYaw);
                             minCameraYaw = min(minCameraYaw, cameraYaw);
                             maxCameraYaw = max(maxCameraYaw, cameraYaw);
                         }
